texlive - Comment on user ownership and prefix.
Since I am increasingly suggesting that anyone wanting to install
the binary might prefer to do it as a normal user, clarify that
root should not do any updates.

At the start of 'setting the PATH' explain why we do things like
this, and mention that we used to bootstrap the source build - that
is to give a clue if a reader decides they want to install ConTeXt
and LuaMetaTeX where I suggest doing a similar thing.