texlive - update comment sabout my test scripts
New version latex-test-20240317 : In the previous version I had
assumed that kpsepath (part of TLCore) would always be present
when looking for TTF and OTF fonts, and failed with an error if
it was missing.

Testing a much smaller binary install yesterday showed it is NOT
always present - the only schemes containing it are medium and
full.  Now fixed, my tests with e.g. basic plus the latexextra
and latexrecommended collections now run.

Also update the version in th comment re testing asymptote.