TexLive 2024 - update binary sizes.
Show minimum (85 MB plain TeX with luatex only for the scripts which
support texlive, an estimate for latex and later engines: I got
2.1 GB with scheme-basic + collection-latexrecommended, or 2.2 GB
with scheme-medium (very different packages), I guess up to 4 GB
for people who make a lot of use of the modern engines.  8.6 GB
for everything.

Show the binary size for ConTeXt in the note on the page for texlive
source - unless somebody wants to maintain ConTeXt I guess that
information need not be updated in future years.