Update to Devel-StackTrace-2.05 (perl dep).
Added explicit boolean overloading for trace objects. Without this,
Perl will use the object's string overloading and then check the
truthiness of the returned string, which is a lot of extra work.
This can produce significant slowdowns in some cases (where large
data is being passed around as a string and then ignored).