Update to HTML-Tagset-3.24 (perl dep).
Added <ins> and <del> to %isPhraseMarkup

Fixed docs about %isBodyElement calling it %isBodyMarkup

Cleaned up metadata and version requirements

No changes have been made to handle HTML5. That is a bigger issue
than "just add the tags to the hashes," because that would change
functionality for existing users who are expecting HTML::Tagset to
be HTML4.