Make several fixes regarding typos and licensing (see below)
Since GLFS contains a significant amount of code from BLFS, for legal
reasons the BLFS MIT license must be in the repo. Instead of relicensing
which is a grey area, this commit sublicenses GLFS, as a downstream of
BLFS which it is. Along with this the changelog was added back as this is
a core part of *LFS maintainers